add dialog

+++ src/app/analysis/rule/rule.component.ts
@@ -1,5 +1,7 @@
 import { Component, OnInit, AfterViewInit } from '@angular/core';
+import { MdDialog, MdDialogRef, MD_DIALOG_DATA } from '@angular/material';
 import { AnalysisService } from '../../services/analysis.service';
+import { DialogConfirmComponent } from '../../component/dialog-confirm/dialog-confirm.component';
 import { ToasterService } from 'angular2-toaster';
 
 declare var $: any;
@@ -15,18 +17,13 @@ export class RuleComponent implements OnInit, AfterViewInit {
   p: number = 1;
   private sortReverse = false;
 
-  constructor(private _analysisService: AnalysisService, private _toasterService: ToasterService) { }
+  constructor(private _analysisService: AnalysisService, private _toasterService: ToasterService, public dialog: MdDialog) { }
 
   async ngOnInit() {
     this.rules = await this._analysisService.getRuleList();
   }
 
   ngAfterViewInit() {
-    // const exampleId: any = $('#example');
-
-    // // exampleId.hide();
-    // console.log(exampleId);
-    // exampleId.DataTable();
   }
 
   onClickOrder(header) {
@@ -35,12 +32,26 @@ export class RuleComponent implements OnInit, AfterViewInit {
   }
 
   async deleteRule(ruleId) {
-    await this._analysisService.deleteRule(ruleId);
-    this._toasterService.pop('success', 'Validation error', 'Formula is required.');
+
+    const dialogRef = this.dialog.open(DialogConfirmComponent, {
+      width: '450px',
+      data: { hint: `Are sure you want to remove rule ${ruleId}?`}
+    });
+
+    dialogRef.afterClosed().subscribe(async (result) => {
+      if (result && result === 'Yes') {
+        await this._analysisService.deleteRule(ruleId);
+        this._toasterService.pop('success', 'Validation error', 'Formula is required.');
+      }
+    });
   }
 
   editRule(ruleId) {
     console.log('edit rule id: ', ruleId);
+    this.openDialog();
   }
 
+  openDialog(): void {
+
+  }
 }
